+++ src/helpers/Gql.php
@@ -38,4 +38,34 @@ class Gql
         }
         return $gqlAccessToken;
     }
+
+    public static function extractGQLFields($parsedQuery)
+    {
+        if (isset($parsedQuery->definitions[0]->selectionSet)) {
+            $selections = iterator_to_array($parsedQuery->definitions[0]->selectionSet->selections);
+            return array_map(function($selection) {
+                $field = [
+                    'field' => $selection->name->value,
+                    'arguments' => [],
+                ];
+                if (isset($selection->arguments)) {
+                    foreach ($selection->arguments as $argument) {
+                        if ($argument->value->kind === 'ListValue') {
+                            $listValues = [];
+                            foreach ($argument->value->values as $listItem) {
+                                $listValues[] = $listItem->value;
+                            }
+                            $field['arguments'][$argument->name->value] = $listValues;
+                        } elseif ($argument->value->kind === 'Variable') {
+                            $field['arguments'][$argument->name->value] = '$' . $argument->value->name->value;
+                        } else {
+                            $field['arguments'][$argument->name->value] = $argument->value->value;
+                        }
+                    }
+                }
+                return $field;
+            }, $selections);
+        }
+        return [];
+    }
 }